:recycle: refactoring: Update deletarNivel method to throw RuntimeException for not found Nivel instead of using ResponseStatusException

+++ src/main/java/com/conectados/conectados/services/NivelService.java
@@ -37,6 +37,10 @@ public class NivelService {
     }
 
     public void deletarNivel(Long id) {
+        Optional<Nivel> nivel = buscarNivelPorId(id);
+        if (nivel.isEmpty()) {
+            throw new RuntimeException("Nível não encontrado");
+        }
         nivelRepository.deleteById(id);
     }
